TNT-40917-1 Update SDK to use new Delivery client after #18 fix (#19)
* Update SDK to use new Delivery client after #18 fix

* Update Delivery client to latest v1.1.1

+++ Source/Adobe.Target.Client/TargetClient.cs
@@ -150,7 +150,7 @@ namespace Adobe.Target.Client
             {
                 if (mbox.Index >= index)
                 {
-                    index = mbox.Index + 1;
+                    index = (int)mbox.Index + 1;
                 }
 
                 existingMboxNames.Add(mbox.Name);
